+++ package.json
@@ -5,30 +5,37 @@
 	"license": "MIT",
 	"scripts": {
 		"start": "webpack-dev-server",
-		"build": "rimraf dist && webpack --mode=production"
+		"build": "rimraf dist && webpack --mode=production",
+		"test": "mocha -r ts-node/register test/*.tsx"
 	},
 	"dependencies": {
 		"@babel/core": "^7.4.3",
 		"@babel/plugin-syntax-dynamic-import": "^7.2.0",
 		"@babel/preset-env": "^7.4.3",
 		"@loadable/component": "^5.7.0",
+		"@types/chai": "^4.1.7",
 		"@types/loadable__component": "^5.6.0",
+		"@types/mocha": "^5.2.6",
 		"@types/react": "^16.8.13",
 		"@types/react-dom": "^16.8.4",
 		"@types/react-router": "^4.4.5",
 		"@types/react-router-dom": "^4.3.2",
+		"@types/react-test-renderer": "^16.8.1",
 		"@types/reactstrap": "^8.0.0",
 		"awesome-typescript-loader": "^5.2.1",
 		"babel-loader": "^8.0.5",
 		"bootstrap": "^4.3.1",
+		"chai": "^4.2.0",
 		"css-loader": "^2.1.1",
 		"html-webpack-plugin": "^3.2.0",
+		"mocha": "^6.1.3",
 		"react": "^16.8.6",
 		"react-dom": "^16.8.6",
 		"react-router": "^5.0.0",
 		"react-router-dom": "^5.0.0",
 		"react-test-renderer": "^16.8.6",
 		"reactstrap": "^8.0.0",
+		"ts-node": "^8.1.0",
 		"typescript": "^3.4.3",
 		"webpack": "^4.30.0",
 		"webpack-cli": "^3.3.0",
