+++ src/components/navigation.tsx
@@ -1,0 +1,46 @@
+import * as React from 'react';
+import Navbar from 'reactstrap/lib/Navbar';
+import NavbarToggler from 'reactstrap/lib/NavbarToggler';
+import Collapse from 'reactstrap/lib/Collapse';
+import Nav from 'reactstrap/lib/Nav';
+import NavItem from 'reactstrap/lib/NavItem';
+import NavLink from 'reactstrap/lib/NavLink';
+
+interface NavigationComponentState {
+	isOpen: boolean;
+}
+export default class NavigationComponent extends React.Component<{}, NavigationComponentState> {
+	constructor(props: any) {
+		super(props);
+		this.toggle = this.toggle.bind(this);
+		this.state = {
+			isOpen: false
+		};
+	}
+
+	toggle() {
+		this.setState({
+			isOpen: !this.state.isOpen
+		});
+	}
+
+	render() {
+		return (
+			<div>
+				<Navbar color="dark" dark expand="md">
+					<NavbarToggler onClick={this.toggle} />
+					<Collapse isOpen={this.state.isOpen} navbar>
+						<Nav className="ml-auto" navbar>
+							<NavItem>
+								<NavLink href="/">Home</NavLink>
+							</NavItem>
+							<NavItem>
+								<NavLink href="/about">About</NavLink>
+							</NavItem>
+						</Nav>
+					</Collapse>
+				</Navbar>
+			</div>
+		);
+	}
+}
